requirements.txt
# requirements for the project
# to quick install all the requirements run the following command
# pip install -r requirements.txt
# if using conda env: conda install -r requirements.txt
# make sure you are in the same directory as the requirements.txt file and that you have activated your virtual environment
#python==3.11.4
# Faker to create random data
Faker==19.6.1
python-dateutil==2.8.2
six==1.16.0

SQLAlchemy==2.0.21

# pyodbc to connecto to the database
pyodbc==4.0.39

uvicorn==0.23.2

# install fastAPI with conda, if you use anaconda "conda install -c conda-forge fastapi"
fastapi==0.103.2

#for password to db
python-dotenv==1.0.0
#anaconda env is env (for Jon to remember)

# for getting API requests and stuff
requests==2.31.0